* Corrijo nombre de interfaces repositorios que no seguían la convención. * Corrijo enrutamiento. - Elimino utils de paginación que no utilizados. - Elimino controlador de Máquinas no utilizado, junto con vista de administración de máquinas. + Implemento borrado lógico para las entidades que lo requerían. + Implemento papeleras para la recuperación de las entidades con borrado lógico.

+++ gestion-limpia/src/main/java/com/burbujas/gestionlimpia/models/services/CajaServiceImpl.java
@@ -90,13 +90,27 @@ public class CajaServiceImpl implements ICajaService{
     }
 
     @Override
+    public List<MovimientoCaja> findAllMovimientosCajaEliminados() {
+        return this.movimientoCajaRepository.findAllMovimientosCajaEliminados();
+    }
+
+    @Override
     public void saveMovimientoCaja(MovimientoCaja movimientoCaja) {
         this.movimientoCajaRepository.save(movimientoCaja);
     }
 
     @Override
     public void deleteMovimientoCaja(MovimientoCaja movimientoCaja) {
-        this.movimientoCajaRepository.delete(movimientoCaja);
+        if(movimientoCaja.getPedido() != null){
+            throw new IllegalArgumentException("No se puede eliminar un movimiento de caja generado por el cobro pedido.");
+        }else{
+            this.movimientoCajaRepository.delete(movimientoCaja);
+        }
+    }
+
+    @Override
+    public void restaurarMovimientoCaja(Long id){
+        this.movimientoCajaRepository.restaurarMovimientoCaja(id);
     }
 
     @Override
